fix: move the provider.go file out of the providers package to prevent circular imports.
